+++ source/generation.h
@@ -1,9 +1,10 @@
 #pragma once
 
 #include "classmap.h"
+#include "metadata.h"
 #include "method_info.h"
 #include "recorder.h"
 
 namespace appmap {
-    std::string generate(const recording &events, bool generate_classmap);
+    std::string generate(const recording &events, bool generate_classmap, const metadata &metadata = {});
 }